+++ pruebas/01-reading-list/bielmc11/src/components/MyBooks.tsx
@@ -9,11 +9,13 @@ import {
 } from "./ui/card";
 import { useAppDispatch, useAppSelector } from "@/hooks/useStore";
 import { Button } from "./ui/button";
-import { removeBook } from "@/store/MyBooks/slice.ts";
+import { removeBook } from "@/store/MyListBooks/slice";
 import { Library } from "@/interfaces/interfaces";
 
 export const MyBooks = () => {
-  const myBooks: Library[] = useAppSelector((state) => state.myBooks as Library[]);
+  const myBooks: Library[] = useAppSelector(
+    (state) => state.myBooks as Library[]
+  );
   const dispatch = useAppDispatch();
 
   return (
